update UI
1. Split voice room and settings section into dedicated windows
2. Use tabs instead of selectables/dropdowns
3. Change wording of button depending on entered room name (private rooms)
4. Make room name/password readonly when currently in a room
5. Add settings button to main window
