+++ DalamudPlugin/ProximityVoiceChat/UI/PluginUIContainer.cs
@@ -11,17 +11,20 @@ public class PluginUIContainer : IDalamudHook
 {
     private readonly IPluginUIPresenter[] pluginUIPresenters;
     private readonly MainWindowPresenter mainWindowPresenter;
+    private readonly ConfigWindowPresenter configWindowPresenter;
     private readonly IDalamudPluginInterface pluginInterface;
     private readonly WindowSystem windowSystem;
 
     public PluginUIContainer(
         IPluginUIPresenter[] pluginUIPresenters,
         MainWindowPresenter mainWindowPresenter,
+        ConfigWindowPresenter configWindowPresenter,
         IDalamudPluginInterface pluginInterface,
         WindowSystem windowSystem)
     {
         this.pluginUIPresenters = pluginUIPresenters ?? throw new ArgumentNullException(nameof(pluginUIPresenters));
         this.mainWindowPresenter = mainWindowPresenter ?? throw new ArgumentNullException(nameof(mainWindowPresenter));
+        this.configWindowPresenter = configWindowPresenter ?? throw new ArgumentNullException(nameof(configWindowPresenter));
         this.pluginInterface = pluginInterface ?? throw new ArgumentNullException(nameof(pluginInterface));
         this.windowSystem = windowSystem ?? throw new ArgumentNullException(nameof(windowSystem));
 
@@ -35,6 +38,7 @@ public class PluginUIContainer : IDalamudHook
     {
         this.pluginInterface.UiBuilder.Draw -= Draw;
         this.pluginInterface.UiBuilder.OpenMainUi -= ShowMainWindow;
+        this.pluginInterface.UiBuilder.OpenConfigUi -= ShowConfigWindow;
         GC.SuppressFinalize(this);
     }
 
@@ -42,6 +46,7 @@ public class PluginUIContainer : IDalamudHook
     {
         this.pluginInterface.UiBuilder.Draw += Draw;
         this.pluginInterface.UiBuilder.OpenMainUi += ShowMainWindow;
+        this.pluginInterface.UiBuilder.OpenConfigUi += ShowConfigWindow;
     }
 
     public void Draw()
@@ -62,5 +67,10 @@ public class PluginUIContainer : IDalamudHook
     private void ShowMainWindow()
     {
         this.mainWindowPresenter.View.Visible = true;
+    }
+
+    private void ShowConfigWindow() 
+    {
+        this.configWindowPresenter.View.Visible = true;
     }
 }
