+++ DalamudPlugin/ProximityVoiceChat/UI/Presenter/MainWindowPresenter.cs
@@ -50,28 +50,6 @@ public class MainWindowPresenter(
 
     private void BindVariables()
     {
-        Bind(this.view.SelectedAudioInputDeviceIndex,
-            b => this.audioDeviceController.AudioRecordingDeviceIndex = b, this.audioDeviceController.AudioRecordingDeviceIndex);
-        Bind(this.view.SelectedAudioOutputDeviceIndex,
-            b => this.audioDeviceController.AudioPlaybackDeviceIndex = b, this.audioDeviceController.AudioPlaybackDeviceIndex);
-        Bind(this.view.PlayingBackMicAudio,
-            b =>
-            {
-                this.audioDeviceController.PlayingBackMicAudio = b;
-                this.voiceRoomManager.PushPlayerAudioState();
-            },
-            this.audioDeviceController.PlayingBackMicAudio);
-        Bind(this.view.PushToTalk,
-            b => 
-            {
-                this.configuration.PushToTalk = b;
-                this.configuration.Save();
-                this.pushToTalkController.UpdateListeners();
-            },
-            this.configuration.PushToTalk);
-        Bind(this.view.SuppressNoise,
-            b => { this.configuration.SuppressNoise = b; this.configuration.Save(); }, this.configuration.SuppressNoise);
-
         Bind(this.view.PublicRoom,
             b => { this.configuration.PublicRoom = b; this.configuration.Save(); }, this.configuration.PublicRoom);
         Bind(this.view.RoomName,
@@ -79,27 +57,6 @@ public class MainWindowPresenter(
         Bind(this.view.RoomPassword,
             s => { this.configuration.RoomPassword = s; this.configuration.Save(); }, this.configuration.RoomPassword);
 
-        Bind(this.view.MasterVolume,
-            f => { this.configuration.MasterVolume = f; this.configuration.Save(); }, this.configuration.MasterVolume);
-        Bind(this.view.AudioFalloffType,
-            t => { this.configuration.FalloffModel.Type = t; this.configuration.Save(); }, this.configuration.FalloffModel.Type);
-        Bind(this.view.AudioFalloffMinimumDistance,
-            f => { this.configuration.FalloffModel.MinimumDistance = f; this.configuration.Save(); }, this.configuration.FalloffModel.MinimumDistance);
-        Bind(this.view.AudioFalloffMaximumDistance,
-            f => { this.configuration.FalloffModel.MaximumDistance = f; this.configuration.Save(); }, this.configuration.FalloffModel.MaximumDistance);
-        Bind(this.view.AudioFalloffFactor,
-            f => { this.configuration.FalloffModel.FalloffFactor = f; this.configuration.Save(); }, this.configuration.FalloffModel.FalloffFactor);
-        Bind(this.view.MuteDeadPlayers,
-            b => { this.configuration.MuteDeadPlayers = b; this.configuration.Save(); }, this.configuration.MuteDeadPlayers);
-        Bind(this.view.MuteDeadPlayersDelayMs,
-            f => { this.configuration.MuteDeadPlayersDelayMs = f; this.configuration.Save(); }, this.configuration.MuteDeadPlayersDelayMs);
-        Bind(this.view.MuteOutOfMapPlayers,
-            b => { this.configuration.MuteOutOfMapPlayers = b; this.configuration.Save(); }, this.configuration.MuteOutOfMapPlayers);
-
-        Bind(this.view.PrintLogsToChat,
-            b => { this.configuration.PrintLogsToChat = b; this.configuration.Save(); }, this.configuration.PrintLogsToChat);
-        Bind(this.view.MinimumVisibleLogLevel,
-            i => { this.configuration.MinimumVisibleLogLevel = i; this.configuration.Save(); }, this.configuration.MinimumVisibleLogLevel);
     }
 
     private void BindActions()
@@ -139,22 +96,6 @@ public class MainWindowPresenter(
 
         this.view.LeaveVoiceRoom.Subscribe(_ => this.voiceRoomManager.LeaveVoiceRoom(false).SafeFireAndForget(ex => this.logger.Error(ex.ToString())));
 
-        this.view.EditingPushToTalkKeybind.Subscribe(b =>
-        {
-            if (b)
-            {
-                this.inputEventSource.SubscribeToKeyDown(OnInputKeyDown);
-            }
-            else
-            {
-                this.inputEventSource.UnsubscribeToKeyDown(OnInputKeyDown);
-            }
-        });
-        this.view.ClearPushToTalkKeybind.Subscribe(_ =>
-        {
-            this.configuration.PushToTalkKeybind = default;
-            this.configuration.Save();
-        });
     }
 
     private void Bind<T>(
@@ -169,18 +110,4 @@ public class MainWindowPresenter(
         reactiveProperty.Subscribe(dataUpdateAction);
     }
 
-    private void OnInputKeyDown(KeyDown k)
-    {
-        this.configuration.PushToTalkKeybind = k.Key;
-        this.configuration.Save();
-        // This callback can be called from a non-framework thread, and UI values should only be modified
-        // on the framework thread (or else the game can crash)
-        this.framework.Run(() =>
-        {
-            if (this.view.EditingPushToTalkKeybind.Value)
-            {
-                this.view.EditingPushToTalkKeybind.Value = false;
-            }
-        });
-    }
 }
