+++ DalamudPlugin/ProximityVoiceChat/UI/View/IMainWindow.cs
@@ -13,26 +13,7 @@ public interface IMainWindow : IPluginUIView
     public IObservable<Unit> JoinVoiceRoom { get; }
     public IObservable<Unit> LeaveVoiceRoom { get; }
 
-    public IReactiveProperty<int> SelectedAudioInputDeviceIndex { get; }
-    public IReactiveProperty<int> SelectedAudioOutputDeviceIndex { get; }
-    public IReactiveProperty<bool> PlayingBackMicAudio { get; }
-    public IReactiveProperty<bool> PushToTalk { get; }
-    public IReactiveProperty<bool> EditingPushToTalkKeybind { get; }
-    public IObservable<Unit> ClearPushToTalkKeybind { get; }
-    public IReactiveProperty<bool> SuppressNoise { get; }
-
     public IObservable<bool> MuteMic { get; }
     public IObservable<bool> Deafen { get; }
 
-    public IReactiveProperty<float> MasterVolume { get; }
-    public IReactiveProperty<AudioFalloffModel.FalloffType> AudioFalloffType { get; }
-    public IReactiveProperty<float> AudioFalloffMinimumDistance { get; }
-    public IReactiveProperty<float> AudioFalloffMaximumDistance { get; }
-    public IReactiveProperty<float> AudioFalloffFactor { get; }
-    public IReactiveProperty<bool> MuteDeadPlayers { get; }
-    public IReactiveProperty<int> MuteDeadPlayersDelayMs { get; }
-    public IReactiveProperty<bool> MuteOutOfMapPlayers { get; }
-
-    public IReactiveProperty<bool> PrintLogsToChat { get; }
-    public IReactiveProperty<int> MinimumVisibleLogLevel { get; }
 }
